+++ lib/bcache.dart
@@ -5,28 +5,28 @@ import 'package:uuid/uuid.dart';
 /// Model
 class Piece {
   String id = Uuid().v1();
-  String prefix = "default";
+  String lot = "default";
   String body;
 
-  Piece({String id, String prefix, String body}) {
+  Piece({String id, String lot, String body}) {
     if (id != null) this.id = id;
-    if (prefix != null) this.prefix = prefix;
+    if (lot != null) this.lot = lot;
     if (body != null) this.body = body;
   }
 
-  factory Piece.fromJson(String prefix, String body) {
-    return Piece(prefix: prefix, body: body);
+  factory Piece.fromJson(String lot, String body) {
+    return Piece(lot: lot, body: body);
   }
 
   Map<String, dynamic> toJson() {
-    return {'id': id, 'prefix': prefix, 'body': body};
+    return {'id': id, 'lot': lot, 'body': body};
   }
 
   /// Implement toString to make it easier to see information about
   /// each piece when using the print statement.
   @override
   String toString() {
-    return 'Piece{id: $id, prefix: $prefix, body: $body}';
+    return 'Piece{id: $id, lot: $lot, body: $body}';
   }
 }
 
@@ -55,7 +55,7 @@ class BCache {
       onCreate: (db, version) {
         // Data types: https://www.sqlite.org/datatype3.html
         return db.execute(
-          'CREATE TABLE $_tableName (id TEXT PRIMARY KEY, prefix TEXT, body TEXT)',
+          'CREATE TABLE $_tableName (id TEXT PRIMARY KEY, lot TEXT, body TEXT)',
         );
       },
       // Set the version. This executes the onCreate function and provides a
@@ -80,25 +80,25 @@ class BCache {
   }
 
   /// @nhancv 10/7/2019: Query data from database
-  Future<List<Piece>> query(String prefix) async {
+  Future<List<Piece>> query(String lot) async {
     // Get a reference to the database.
     final Database db = await database;
 
     final List<Map<String, dynamic>> results =
-        await db.query(_tableName, where: "prefix = ?", whereArgs: [prefix]);
+        await db.query(_tableName, where: "lot = ?", whereArgs: [lot]);
 
     // Convert the List<Map<String, dynamic> into a List<Piece>.
     return List.generate(results.length, (i) {
       return Piece(
         id: results[i]['id'],
-        prefix: results[i]['prefix'],
+        lot: results[i]['lot'],
         body: results[i]['body'],
       );
     });
   }
 
-  /// @nhancv 10/7/2019: Query data from database
-  Future<Piece> queryById(String id) async {
+  /// @nhancv 10/7/2019: Query data by id from database
+  Future<Piece> queryId(String id) async {
     // Get a reference to the database.
     final Database db = await database;
 
@@ -108,9 +108,26 @@ class BCache {
     if (results.length == 0) return null;
     return Piece(
       id: results[0]['id'],
-      prefix: results[0]['prefix'],
+      lot: results[0]['lot'],
       body: results[0]['body'],
     );
+  }
+
+  /// @nhancv 10/7/2019: Query data from database
+  Future<List<Piece>> queryAll() async {
+    // Get a reference to the database.
+    final Database db = await database;
+
+    final List<Map<String, dynamic>> results = await db.query(_tableName);
+
+    // Convert the List<Map<String, dynamic> into a List<Piece>.
+    return List.generate(results.length, (i) {
+      return Piece(
+        id: results[i]['id'],
+        lot: results[i]['lot'],
+        body: results[i]['body'],
+      );
+    });
   }
 
   /// @nhancv 10/7/2019: Update data
@@ -140,6 +157,28 @@ class BCache {
     );
   }
 
+  /// @nhancv 10/9/2019: Delete data in lot
+  Future<void> deleteLot(String lot) async {
+    // Get a reference to the database.
+    final db = await database;
+
+    // Remove the data from the database.
+    await db.delete(
+      _tableName,
+      where: "lot = ?",
+      whereArgs: [lot],
+    );
+  }
+
+  /// @nhancv 10/9/2019: Delete all
+  Future<void> deleteAll() async {
+    // Get a reference to the database.
+    final db = await database;
+
+    // Remove the data from the database.
+    await db.delete(_tableName);
+  }
+
   /// @nhancv 10/7/2019: Execute SQL query
   Future<void> execute(String sql, [List<dynamic> arguments]) async {
     // Get a reference to the database.
@@ -151,12 +190,14 @@ class BCache {
 }
 
 /// Example
-//var piece = Piece(
-//  body: "body1",
-//);
-//
 //BCache bCache = BCache();
 //await bCache.init();
+//
+//// Piece instance
+//var piece = Piece(
+//  lot: "default",
+//  body: "body1",
+//);
 //
 //// Insert a piece into the database.
 //await bCache.insert(piece);
@@ -175,4 +216,29 @@ class BCache {
 //await bCache.delete(piece.id);
 //
 //// Print the list of pieces (empty).
+//print(await bCache.query("default"));
+//
+//// piece2 instance
+//var piece2 = Piece(
+//  lot: "lot2",
+//  body: "body1",
+//);
+//
+//// Insert a piece and piece2 into the database.
+//await bCache.insert(piece);
+//await bCache.insert(piece2);
+//
+//// Print all items (default and lot2)
+//print(await bCache.queryAll());
+//
+//// Delete lot2
+//await bCache.deleteLot("lot2");
+//
+//// Print all items (only default lot)
+//print(await bCache.queryAll());
+//
+//// Delete all
+//await bCache.deleteAll();
+//
+//// Print lot2 items
 //print(await bCache.query("default"));